pyproject.toml
[tool.poetry]
name = "lyndon-exa-data"
version = "0.1.0"
description = ""
authors = ["Lyndon Fan <lyndon0808@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pandas = "^2.1.2"
"fhir.resources" = "^7.0.2"
orjson = "^3.9.10"
sqlalchemy = "^2.0.23"
fsspec = "^2023.10.0"
pymongo = "^4.6.0"
numpy = "^1.26.1"
psycopg2-binary = "^2.9.9"
sqlalchemy-cockroachdb = "^2.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
ruff = "^0.1.4"
python-dotenv = "^1.0.0"
pytest-mock = "^3.12.0"
ipython = "^8.17.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
[tool.ruff.lint.extend-per-file-ignores]
"__init__.py" = ["E402", "F401"]